+++ ergo_std/ChangeLog.md
@@ -22,6 +22,8 @@
   option to make it _only_ serve as a runtime cache.
 * Add `pattern` option to `error:throw` to indicate the error is a pattern
   error (which may be ignored by something like `match`).
+* `value:meta:get` and `env:get` return `Unset` rather than `()` if a return is
+  not available.
 
 ### Bugfixes
 * Fix an issue where `match` incorrectly detected bind errors when a binding in